requirements.txt
agate==1.7.0
alembic==1.10.4
aniso8601==9.0.1
anyio==3.6.2
asn1crypto==1.5.1
attrs==23.1.0
awscli==1.27.142
Babel==2.12.1
backoff==2.2.1
black==23.3.0
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
coloredlogs==14.0
croniter==1.3.15
cryptography==40.0.2
dagit==1.3.6
dagster==1.3.6
dagster-graphql==1.3.6
dbt-core==1.5.0
dbt-extractor==0.4.1
dbt-snowflake==1.5.1
docstring-parser==0.15
docutils==0.16
Faker==18.9.0
filelock==3.12.0
fsspec==2023.5.0
future==0.18.3
gql==3.4.1
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.47.5
grpcio-health-checking==1.47.5
h11==0.14.0
hologram==0.0.16
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
isodate==0.6.1
jaraco.classes==3.2.3
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
keyring==23.13.1
leather==0.3.4
Logbook==1.5.3
Mako==1.2.4
MarkupSafe==2.1.2
mashumaro==3.6
minimal-snowplow-tracker==0.0.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==2.8.8
numpy==1.24.3
oscrypto==1.3.0
packaging==23.1
pandas==2.0.1
parsedatetime==2.4
pathspec==0.11.1
pendulum==2.1.2
platformdirs==3.5.1
protobuf==4.23.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.8
PyJWT==2.7.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.7.2
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.0.3
SQLAlchemy==1.4.48
sqlparse==0.4.3
starlette==0.27.0
tabulate==0.9.0
text-unidecode==1.3
tomli==2.0.1
toposort==1.10
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
universal_pathlib==0.0.23
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
